functions/fun_one/requirements.txt
functions-framework==3.*
google-auth==2.6.2
